% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusMinus.R
\name{minus}
\alias{minus}
\title{Minus (-)}
\usage{
minus(x, y)
}
\arguments{
\item{x}{The base of the subtraction (from which the other integer is subtracted).}

\item{y}{The subtraction from the base.}
}
\value{
The difference between x and y.
}
\description{
Subtracts an integer from another integer (x-y).
}
\details{
Calls the internal function EVAL with \emph{x} and \emph{-y}. EVAL implements a loop of unit value addition.
}
\examples{
minus(1, 10)
minus(-3, -3)
minus(-9, 1)
minus(1, -1)
}
\seealso{
\code{while}
}
\author{
Istvan Andras Horvath
}
