% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusMinus.R
\name{plus}
\alias{plus}
\title{Plus (+)}
\usage{
plus(x, y)
}
\arguments{
\item{x}{The base of the addition (to which the other integer is added).}

\item{y}{The addition to the base.}
}
\value{
The sum of x and y.
}
\description{
Adds an integer to another integer (x+y).
}
\details{
Calls the internal function EVAL with \emph{x} and \emph{y}. EVAL implements a loop of unit value addition.
}
\examples{
plus(7, 5)
plus(9, -3)
plus(-6, 4)
plus(-8, -8)
}
\seealso{
\code{while}
}
\author{
Istvan Andras Horvath
}
